webapi/tct-gumallow-w3c-tests/gumallow/navigator_audio_getting_datastream.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Cui,Jieqiong <jieqiongx.cui@intel.com>
        Xinx, liu <xinx.liu@intel.com>

-->

<html>
  <head>
    <title>GetUserMedia Test: navigator_audio_getting_datastream</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://dev.w3.org/2011/webrtc/editor/getusermedia-20111130.html#navigatorusermedia" />
    <meta name="flags" content="" />
    <meta name="assert" content="we can access to the user's microphone by calling the API 'getUserMedia'." />
    <script src="../resources/testharness.js"></script>
    <script src="../resources/testharnessreport.js"></script>
    <script type="text/javascript">
        navigator.getUserMedia = navigator.getUserMedia || navigator.webkitGetUserMedia;
        function gotUserMediaSuccess(s) {
            var url =!!webkitURL ? webkitURL.createObjectURL(s):window.URL.createObjectURL(s);
            document.getElementById("localView").src = url;
            assert_true(true, "getUserMedia SuccessCallback tests");
        }
        function gotUserMediaError(error) {
            assert_true(false, "It should enter gotUserMediaSuccess.");
        }
        function testUserMedia() {
            try {
                if (navigator.getUserMedia) {
                    navigator.getUserMedia({audio: true}, gotUserMediaSuccess, gotUserMediaError);
                } else {
                    assert_true(false, "No microphone installed in your computer");
                }
            } catch (e) {
                assert_true(false, "No microphone installed in your computer");
            }
        }
        test(testUserMedia, "Test Description: Test GetUserMedia SuccessCallback.");
    </script>
  </head>
  <body>
    <div id="log"></div>
    <audio id="localView" autoplay></audio>
  </body>
</html>
